backend/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# 任务队列
celery==5.3.6
flower==2.0.1

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]

# AI SDK
anthropic==0.18.1
openai==1.10.0

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.1

# 文件处理
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.2.0
python-magic==0.4.27

# 向量数据库（可选，根据选择使用）
# pgvector (PostgreSQL 扩展，需在数据库安装)
# qdrant-client==1.7.0

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# 类型提示
types-redis==4.6.0.20240125
